requirements.txt
APScheduler==3.6.1
asn1crypto==1.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
certifi==2019.9.11
cffi==1.13.1
chainer==6.5.0
chardet==3.0.4
Click==7.0
cryptography==2.8
cycler==0.10.0
dateutils==0.6.7
dropbox==9.4.0
filelock==3.0.12
Flask==1.1.1
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
line-bot-sdk==1.14.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.17.3
oauthlib==3.1.0
paramiko==2.6.0
parse==1.12.1
pickle5==0.0.9
Pillow==6.2.1
pip==19.3.1
protobuf==3.10.0
pyasn1==0.4.7
pycparser==2.19
PyNaCl==1.3.0
pyparsing==2.4.2
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
setuptools==41.6.0
six==1.12.0
soupsieve==1.9.5
tweepy==3.8.0
typing==3.7.4.1
typing-extensions==3.7.4.1
tzlocal==2.0.0
urllib3==1.25.6
Werkzeug==0.16.0
zlib-wrapper==0.1.3
